Refactored time stuff out of IchnosCF
